# Make Data.Sequence.fromList more eager